+++ SpringBanque-Data/src/main/java/fr/afcepf/al28/seb/banque/data/api/IDaoCompte.java
@@ -1,9 +1,10 @@
-package fr.afcepf.al28.idao;
+package fr.afcepf.al28.seb.banque.data.api;
 
 import fr.afcepf.al28.seb.banque.entity.Compte;
 
 public interface IDaoCompte {
 
-	public Compte addCompte();
+    Compte creerCompte(Compte compte);
+    
 	
 }
